pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "formula-ai"
version = "2.0.0"
description = "AI智能文档排版工具"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Waterdrop", email = "waterdrop26651@gmail.com"}
]
keywords = ["document", "formatting", "ai", "word", "docx"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Text Processing",
    "Topic :: Office/Business",
]

dependencies = [
    "python-docx>=0.8.11",
    "requests>=2.28.1",
    "pillow>=9.3.0",
    "chardet>=5.0.0",
    "json5>=0.9.10",
    "streamlit>=1.28.0",
]

[project.optional-dependencies]
gui = [
    "PyQt6>=6.4.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/waterdrop26651/FormulaAI"
Repository = "https://github.com/waterdrop26651/FormulaAI"

[project.scripts]
formulaai-eval = "src.runtime.eval_harness:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
